+++ packages/domain-event/src/services/geocoding/index.ts
@@ -1,0 +1,2 @@
+export * from './geocoding.service.js';
+export * from './strategy/index.js';