+++ packages/shared/src/types/event-location.ts
@@ -1,0 +1,4 @@
+export interface EventLocation {
+  lat: number;
+  long: number;
+}