+++ packages/errors-nest/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.13.0
+
+### Minor Changes
+
+- geolocode event
+
 ## 0.12.0
 
 ### Minor Changes